replacing typings with new typescript 2 flavour
