+++ package.json
@@ -38,6 +38,17 @@
     "webrx": "^1.5.0"
   },
   "devDependencies": {
+    "@types/chai": "^3.4.34",
+    "@types/classnames": "0.0.31",
+    "@types/ix.js": "^1.0.17",
+    "@types/mocha": "^2.2.32",
+    "@types/node": "^6.0.46",
+    "@types/react": "^0.14.47",
+    "@types/react-addons-css-transition-group": "^0.14.17",
+    "@types/react-addons-transition-group": "^0.14.17",
+    "@types/react-bootstrap": "0.0.35",
+    "@types/react-dom": "^0.14.18",
+    "@types/react-fa": "^4.0.27",
     "babel-core": "^6.18.2",
     "babel-preset-es2015": "^6.18.0",
     "babel-register": "^6.18.0",
