+++ src/types/jquery-deparam/index.d.ts
@@ -1,0 +1,6 @@
+declare module 'jquery-deparam' {
+  function deparam<T>(params: string, coerce?: boolean): T;
+
+  namespace deparam {}
+  export = deparam;
+}
