+++ src/types/jquery-param/index.d.ts
@@ -1,0 +1,6 @@
+declare module 'jquery-param' {
+  function param(obj: any): string;
+
+  namespace param {}
+  export = param;
+}
