+++ src/types/webrx/index.d.ts
@@ -1,0 +1,3 @@
+declare module 'webrx' {
+  export = wx;
+}
