--- src/typings/jquery-param/jquery-param.d.ts
@@ -1,10 +1,0 @@
-declare namespace __JQueryParam {
-  interface JQueryParamStatic {
-    (obj: any): string;
-  }
-}
-
-declare module 'jquery-param' {
-    const param: __JQueryParam.JQueryParamStatic;
-    export = param;
-}
