--- src/typings/react-css-transition-replace/react-css-transition-replace.d.ts
@@ -1,6 +1,0 @@
-/// <reference path="../../../typings/main/ambient/react-addons-css-transition-group/index.d.ts" />
-
-declare module 'react-css-transition-replace' {
-  const CSSTransitionReplace: __React.CSSTransitionGroup;
-  export = CSSTransitionReplace;
-}
